Fix unit tests and stubs for design consistency
Align tests and stubs with design docs after consistency audit:

- MidasConfig: fix defaults to match design 03-05 (multiplier_mode=
  "adaptive", cool_down: float=0.05, mult_max=5.0, er_target=0.1)
- AdaptiveMultiplier: fix cool_down type from int to float (deflation
  rate, not episode skip count); add 5-zone bang-bang controller tests
  (dead zone, moderate/strong inflate, emergency double, transitions)
- TrainingLog: add from_entity/balance_after assertions for allocate,
  consume, transfer records; add is_active edge case tests
- Observer: add missing on_time_paused/on_time_resumed tests (7/7 hooks)
- ConfigSnapshotStore: add min_eta, top_k, combined filter tests
- Integration tests: fix S_w formula from S_exec*(1-β*(1-S_llm)) to
  S_exec+(1-S_exec)*β*S_llm per design 05-01

+++ midas_agent/scheduler/budget_allocator.py
@@ -8,7 +8,7 @@ class AdaptiveMultiplier:
         mode: str,
         init_value: float,
         er_target: float | None = None,
-        cool_down: int | None = None,
+        cool_down: float | None = None,
         mult_min: float | None = None,
         mult_max: float | None = None,
     ) -> None:
